+++ src/components/AddLinkForm/AddLinkForm.tsx
@@ -1,11 +1,12 @@
 import "./AddLinkForm.css";
+import React from "react";
 import LinkIcon from "../../assets/icon/link.png";
 import { useContext } from "react";
 import ModalContext from "../Modal/ModalContext";
 
 function AddLinkForm() {
-  const { openModal, setModalType } = useContext(ModalContext);
-  const onClickModalOpen = (e) => {
+  const { openModal, setModalType } = useContext(ModalContext)!;
+  const onClickModalOpen = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
     openModal();
     setModalType("folderAdd");
