+++ src/components/Card/EditableStarButton.tsx
@@ -1,12 +1,13 @@
+import React from "react";
 import starIcon from "../../assets/svg/star.svg";
 import "./EditableStarButton.css";
 
 function EditableStarButton() {
-    return (
-        <button className="editable-star">
-            <img src={starIcon} alt="즐겨찾기" />
-        </button>
-    );
+  return (
+    <button className="editable-star">
+      <img src={starIcon} alt="즐겨찾기" />
+    </button>
+  );
 }
 
 export default EditableStarButton;
