--- src/pages/NotFoundPage.jsx
@@ -1,5 +1,0 @@
-const NotFoundPage = () => {
-    return <div>페이지 없음</div>;
-};
-
-export default NotFoundPage;
