+++ src/pages/NotFoundPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const NotFoundPage = () => {
+  return <div>페이지 없음</div>;
+};
+
+export default NotFoundPage;
